signal_fir: the reverse sweep of rad reads tapes only (exp, exp10, sqrt load their taped output, abs rebuilds |x| from its taped operand) and refuses a value it would have to recompute with state, so a recursion held by such an operand no longer advances its forward state once more per block; cranelift-ffi: a slider's default reaches a -double program as the source's double, not its f32 rounding
Both found by faust-diff-amp (the DDSP guitar amp of Yeh et al. 2024 ported next to faust-rs). The loss lane of a rad over a gated recurrent unit differed between block sizes from the first frame after a block boundary: propagate_bra_unary_math_adj loaded the operand of every unary rule, the tape analysis tapes the output of exp/exp10/sqrt rather than the operand, and load_bra_fwd_value fell back to lower_signal in the reverse loop, emitting the operand's computation with the state update of the recursion inside it (a phasor, a noise source). abs had the mirror case (its output loaded, not taped). Tests: a_rad_whose_exp_operand_holds_a_recursion_keeps_the_primal_across_blocks, a_rad_whose_abs_operand_holds_a_recursion_keeps_the_primal_across_blocks (two blocks of 64 equal frames 64..128 of one block of 128), the guard reported in FRS-SFIR-0004. The runtime descriptor kept control defaults as f32 and widened them into f64 zones: hslider("lr", 0.02, ...) started at 0.019999999552965164 where --set lr=0.02 delivered 0.02; test a_slider_starts_from_the_double_its_source_wrote. Journal 2026-09-20.

+++ crates/cranelift-ffi/src/instance.rs
@@ -848,20 +848,22 @@ fn apply_control_defaults(
         };
         match &field.kind {
             StructFieldKind::Scalar(FirType::Float32) => {
-                write_f32(dsp_state, field.offset_bytes as usize, *value);
+                write_f32(dsp_state, field.offset_bytes as usize, *value as f32);
             }
             // A `FaustFloat` control zone resolves to f32 or f64 depending on the
             // precision the backend compiled with; write it at the field's actual
-            // width so a `-double` zone (8 bytes) is not corrupted by an f32 store.
+            // width so a `-double` zone (8 bytes) is not corrupted by an f32 store,
+            // and with the exact double so that a default and a `--set` of the
+            // same value are the same bits.
             StructFieldKind::Scalar(FirType::FaustFloat) => {
                 if field.size_bytes >= 8 {
-                    write_f64(dsp_state, field.offset_bytes as usize, f64::from(*value));
+                    write_f64(dsp_state, field.offset_bytes as usize, *value);
                 } else {
-                    write_f32(dsp_state, field.offset_bytes as usize, *value);
+                    write_f32(dsp_state, field.offset_bytes as usize, *value as f32);
                 }
             }
             StructFieldKind::Scalar(FirType::Float64) => {
-                write_f64(dsp_state, field.offset_bytes as usize, f64::from(*value));
+                write_f64(dsp_state, field.offset_bytes as usize, *value);
             }
             StructFieldKind::Scalar(FirType::Int32) => {
                 write_i32(dsp_state, field.offset_bytes as usize, *value as i32);
